Fixes and test updates.
