+++ src/OpenTelemetry/Internal/EnvironmentVariables/EnvironmentVariablesExtensions.cs
@@ -4,49 +4,24 @@
 
 #nullable enable
 
-using System;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Configuration.EnvironmentVariables;
 
-namespace Microsoft.Extensions.Configuration
+namespace OpenTelemetry.Internal;
+
+/// <summary>
+/// Extension methods for registering <see cref="EnvironmentVariablesConfigurationProvider"/> with <see cref="IConfigurationBuilder"/>.
+/// </summary>
+internal static class EnvironmentVariablesExtensions
 {
     /// <summary>
-    /// Extension methods for registering <see cref="EnvironmentVariablesConfigurationProvider"/> with <see cref="IConfigurationBuilder"/>.
+    /// Adds an <see cref="IConfigurationProvider"/> that reads configuration values from environment variables.
     /// </summary>
-    internal static class EnvironmentVariablesExtensions
+    /// <param name="configurationBuilder">The <see cref="IConfigurationBuilder"/> to add to.</param>
+    /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
+    public static IConfigurationBuilder AddEnvironmentVariables(this IConfigurationBuilder configurationBuilder)
     {
-        /// <summary>
-        /// Adds an <see cref="IConfigurationProvider"/> that reads configuration values from environment variables.
-        /// </summary>
-        /// <param name="configurationBuilder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-        /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-        public static IConfigurationBuilder AddEnvironmentVariables(this IConfigurationBuilder configurationBuilder)
-        {
-            configurationBuilder.Add(new EnvironmentVariablesConfigurationSource());
-            return configurationBuilder;
-        }
-
-        /// <summary>
-        /// Adds an <see cref="IConfigurationProvider"/> that reads configuration values from environment variables
-        /// with a specified prefix.
-        /// </summary>
-        /// <param name="configurationBuilder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-        /// <param name="prefix">The prefix that environment variable names must start with. The prefix will be removed from the environment variable names.</param>
-        /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-        public static IConfigurationBuilder AddEnvironmentVariables(
-            this IConfigurationBuilder configurationBuilder,
-            string? prefix)
-        {
-            configurationBuilder.Add(new EnvironmentVariablesConfigurationSource { Prefix = prefix });
-            return configurationBuilder;
-        }
-
-        /// <summary>
-        /// Adds an <see cref="IConfigurationProvider"/> that reads configuration values from environment variables.
-        /// </summary>
-        /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-        /// <param name="configureSource">Configures the source.</param>
-        /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-        public static IConfigurationBuilder AddEnvironmentVariables(this IConfigurationBuilder builder, Action<EnvironmentVariablesConfigurationSource>? configureSource)
-            => builder.Add(configureSource);
+        configurationBuilder.Add(new EnvironmentVariablesConfigurationSource());
+        return configurationBuilder;
     }
 }
